Have Travis install OpenMC during install step and have script call ctest
