+++ .travis.yml
@@ -18,18 +18,17 @@ env:
   global:
     - FC=gfortran
     - MPI_DIR=/usr
-    - PHDF5_DIR=/usr
-    - HDF5_DIR=/usr
+    - HDF5_ROOT=/usr
     - OMP_NUM_THREADS=2
     - OPENMC_CROSS_SECTIONS=$HOME/nndc_hdf5/cross_sections.xml
     - OPENMC_ENDF_DATA=$HOME/endf-b-vii.1
     - OPENMC_MULTIPOLE_LIBRARY=$HOME/multipole_lib
     - PATH=$PATH:$HOME/NJOY2016/build
   matrix:
-    - OPENMC_CONFIG="^hdf5-debug$"
-    - OPENMC_CONFIG="^omp-hdf5-debug$"
-    - OPENMC_CONFIG="^mpi-hdf5-debug$"
-    - OPENMC_CONFIG="^phdf5-debug$"
+    - OMP=n MPI=n PHDF5=n
+    - OMP=y MPI=n PHDF5=n
+    - OMP=n MPI=y PHDF5=n
+    - OMP=n MPI=y PHDF5=y
 
 before_install:
   - sudo add-apt-repository ppa:nschloe/hdf5-backports -y
