+++ tools/ci/travis-install.sh
@@ -15,5 +15,8 @@ if [[ "$TRAVIS_PYTHON_VERSION" == "3.4" ]]; then
     pip install pandas==0.20.3
 fi
 
+# Build and install
+python tools/ci/travis-install.py
+
 # Install OpenMC in editable mode
 pip install -e .[test]
